requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Kafka - Use kafka-python-ng for Python 3.12+ compatibility
kafka-python-ng==2.2.2

# Data Processing
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
scipy>=1.11.0,<2.0.0

# Database
psycopg2-binary==2.9.9
influxdb-client==1.38.0
sqlalchemy==2.0.23

# Machine Learning (for later phases)
scikit-learn>=1.3.0,<2.0.0

# API & Dashboard (for later phases)
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
websockets==12.0

# Monitoring & Logging
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
jupyter==1.0.0
matplotlib>=3.8.0,<4.0.0
<<<<<<< HEAD
seaborn>=0.13.0,<1.0.0
wfdb==4.1.2
=======
seaborn>=0.13.0,<1.0.0
>>>>>>> 5518597 (Initial commit: reset and push to master)
